+++ ibtk/src/solvers/impls/CCPoissonHypreLevelSolver.cpp
@@ -515,7 +515,7 @@ void CCPoissonHypreLevelSolver::allocateHypreData()
 void CCPoissonHypreLevelSolver::setMatrixCoefficients_aligned()
 {
     // Set matrix entries and copy them to the hypre matrix structures.
-    const int stencil_sz = d_stencil_offsets.size();
+    const int stencil_sz = static_cast<int>(d_stencil_offsets.size());
     std::vector<int> stencil_indices(stencil_sz);
     for (int i = 0; i < stencil_sz; ++i)
     {
