+++ ibtk/src/solvers/impls/PETScLevelSolver.cpp
@@ -367,7 +367,7 @@ void PETScLevelSolver::setupNullspace()
     ierr = MatNullSpaceCreate(
         PETSC_COMM_WORLD,
         d_nullspace_contains_constant_vec ? PETSC_TRUE : PETSC_FALSE,
-        petsc_nullspace_basis_vecs.size(),
+        static_cast<int>(petsc_nullspace_basis_vecs.size()),
         (petsc_nullspace_basis_vecs.empty() ? NULL : &petsc_nullspace_basis_vecs[0]),
         &d_petsc_nullsp);
     IBTK_CHKERRQ(ierr);
